\name{SpecDataDolphin}
\alias{SpecDataDolphin}
\alias{MetaDataDolphin}
\docType{data}
\title{Package data.}
\description{Metadata and spectral data for SpecLibDolphin2016. Called by \code{LibraryReport()}.}
\usage{SpecDataDolphin}
\format{\code{SpecDataDolphin} is a data frame containing the mass spectra.  \code{MetaDataDolphin} is a data frame containing the metadata.
}
\keyword{datasets} 

