\name{SpecLibDolphin2016-package}
\alias{SpecLibDolphin2016-package}
\docType{package}
\title{
Mass Spectral Library Brazil Bottlenose Dolphin Blubber 2016
}
\description{
Contains the code and mass spectral library described in: Non-targeted screening of halogenated organic compounds in bottlenose dolphins (Tursiops truncatus) from Rio de Janeiro, Brazil.

The corresponding article includes a description of
the study goals, samples, and GCxGC/TOF-MS
analytical method. The original PDF version of this library is supplementary
material to the article.
}
\keyword{ package }
